+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-## 0.0.3
+## v0.1.0
 
-* Fixed the bug to make the project run on all platforms.
+Multiple changes:
+  - `printRow` is now `printKeyValue` (possible breaking change for existing users of original package)
+  - Added `printFlex` function that allows printing of multiple items from a list along with user-specified width and alignment
+  - Added config options (such as `printWidth`, `leftPadding`, 'rightPadding`)
+  - Added more config options to the `printText` func, such as (alignment, bold, title)
+  - Added fully functional `example`
+  - Updated .gitignore